Do not report non-fatal errors immediately
The main motivation for such change is the fact we can report several non-fatal errors in a small time span. For example, imagine successful response from the server, which contains a list of records. Some of those records are invalid and we want to report non-fatal errors for such records.